+++ packages/liveblocks-react-ui/src/_private/index.ts
@@ -1,6 +1,14 @@
 // Private APIs
 
-export { Button } from "../components/internal/Button";
+export { Button, SelectButton } from "../components/internal/Button";
 export { List } from "../components/internal/List";
+export {
+  ShortcutTooltip,
+  Tooltip,
+  TooltipProvider,
+} from "../components/internal/Tooltip";
 export { User } from "../components/internal/User";
 export * from "../icons";
+export { capitalize } from "../utils/capitalize";
+export { useInitial } from "../utils/use-initial";
+export { useRefs } from "../utils/use-refs";