+++ starter-kits/nextjs-starter-kit/components/TextEditor/ToolbarAlignment.tsx
@@ -1,3 +1,4 @@
+import { Toolbar } from "@liveblocks/react-tiptap";
 import { Editor } from "@tiptap/react";
 import {
   AlignCenterIcon,
@@ -5,8 +6,6 @@ import {
   AlignLeftIcon,
   AlignRightIcon,
 } from "@/icons";
-import { Button } from "@/primitives/Button";
-import styles from "./Toolbar.module.css";
 
 type Props = {
   editor: Editor | null;
@@ -15,57 +14,37 @@ type Props = {
 export function ToolbarAlignment({ editor }: Props) {
   return (
     <>
-      <Button
-        variant="subtle"
-        className={styles.toolbarButton}
+      <Toolbar.Toggle
+        name="Align left"
+        icon={<AlignLeftIcon />}
+        active={editor?.isActive({ textAlign: "left" }) ?? false}
         onClick={() => editor?.chain().focus().setTextAlign("left").run()}
         disabled={!editor?.can().chain().focus().setTextAlign("left").run()}
-        data-active={
-          editor?.isActive({ textAlign: "left" }) ? "is-active" : undefined
-        }
-        aria-label="Align left"
-      >
-        <AlignLeftIcon />
-      </Button>
+      />
 
-      <Button
-        variant="subtle"
-        className={styles.toolbarButton}
+      <Toolbar.Toggle
+        name="Align center"
+        icon={<AlignCenterIcon />}
+        active={editor?.isActive({ textAlign: "center" }) ?? false}
         onClick={() => editor?.chain().focus().setTextAlign("center").run()}
         disabled={!editor?.can().chain().focus().setTextAlign("center").run()}
-        data-active={
-          editor?.isActive({ textAlign: "center" }) ? "is-active" : undefined
-        }
-        aria-label="Align center"
-      >
-        <AlignCenterIcon />
-      </Button>
+      />
 
-      <Button
-        variant="subtle"
-        className={styles.toolbarButton}
+      <Toolbar.Toggle
+        name="Align right"
+        icon={<AlignRightIcon />}
+        active={editor?.isActive({ textAlign: "right" }) ?? false}
         onClick={() => editor?.chain().focus().setTextAlign("right").run()}
         disabled={!editor?.can().chain().focus().setTextAlign("right").run()}
-        data-active={
-          editor?.isActive({ textAlign: "right" }) ? "is-active" : undefined
-        }
-        aria-label="Align right"
-      >
-        <AlignRightIcon />
-      </Button>
+      />
 
-      <Button
-        variant="subtle"
-        className={styles.toolbarButton}
+      <Toolbar.Toggle
+        name="Justify"
+        icon={<AlignJustifyIcon />}
+        active={editor?.isActive({ textAlign: "justify" }) ?? false}
         onClick={() => editor?.chain().focus().setTextAlign("justify").run()}
         disabled={!editor?.can().chain().focus().setTextAlign("justify").run()}
-        data-active={
-          editor?.isActive({ textAlign: "justify" }) ? "is-active" : undefined
-        }
-        aria-label="Justify"
-      >
-        <AlignJustifyIcon />
-      </Button>
+      />
     </>
   );
 }